chore: refacor to move journey components & fix tramslations with embeded links.
