--- ui/src/components/Absentee/index.js
@@ -1,2 +1,0 @@
-import Absentee from './Absentee';
-export default Absentee;
