--- ui/src/components/GetInformed/index.js
@@ -1,2 +1,0 @@
-import GetInformed from './GetInformed';
-export default GetInformed;
