--- ui/src/components/Register/index.js
@@ -1,2 +1,0 @@
-import Register from './Register';
-export default Register;
