+++ ui/src/plugins/en.json
@@ -71,15 +71,14 @@
   "registrationDeadline": "Registration deadline is",
   "registerLiveInUsTerritory": "Do you live is a US territory? Check with your local election officials about voting.",
   "registerTerritoryCitizensRights": "Residents of U.S. territories do not have voting representation in the United States Congress",
-  "registerOnline": "Register online",
-  "registerMail": "Register by mail",
+  "registerLinks": "[Register online]({onlineUrl}) or [Register by mail]({mailUrl}).",
   "registerElectionDay": "Can I register to vote on election day?",
   "registerOnlineQ": "Can I register online?",
   "registerAlready": "Already registered?",
 
   "absenteeTitle": "Can't be there? Let us help you connect",
   "absenteeDesc": "Absentee voting is as easy as signing up before your state's deadline.",
-  "absenteeRequest": "Request an absentee ballot",
+  "absenteeRequest": "Request an absentee ballot [online]({url}).",
   "absenteeImageAlt": "stock image of voters at a voting poll",
   "absenteeRequested": "Already requested ballot or planning to vote early?",
   "absenteeMoreInfo": "[More information]({moreUrl}) from your state about absentee ballots.",
@@ -105,6 +104,7 @@
   "getInformedCongFind": "Find your congressional district",
   "getInformedTopContests": "These are the top contests in your state.",
   "getInformedMoreInfo": "More information about candidates",
+  "getInformedCandidateOrder": "The order of candidates listed below is determined by local election officials.",
 
   "voteTitle": "Vote early or find your polling location",
   "voteTitleDesc": "Need help knowing where to go on voting day? Want to know what location in your precinct allows for early voting? {0} can help you figure it out. Input your address below, and select either \"Early Voting\" or \"Polling Location\". We'll do the rest.",
