+++ ui/src/views/JourneyPage/_absentee.scss
