+++ ui/src/views/JourneyPage/_get-informed.scss
@@ -1,5 +1,5 @@
 @import '@/styles/_theme.scss';
-@import 'carbon-components/scss/components/accordion/accordion';
+@import '../../../node_modules/carbon-components/scss/components/accordion/accordion';
 
 .candidate {
   @include type-style('body-short-02');
