+++ ui/src/views/JourneyPage/_register.scss
@@ -27,7 +27,6 @@
 
     .cv-link {
       @include type-style('body-short-02');
-      margin-left: 1rem;
     }
     &-deadline {
       @include type-style('productive-heading-03');
